requirements.txt
tensorflow==2.12.0
absl-py>=1.1.0
pygame==2.1.0
gym==0.25.0
gym[Box2D]
matplotlib==3.5.2
numpy==1.23.1
tqdm>=4.64.0
